requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.79.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
python-dotenv==1.1.0
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
